script to sync files to bmc
Summary:
Simple script to sync fixmybmc files to the correct location on a BMC so that it can be run without needing to rebuild an image. This is useful for developers and also anyone who wants to use fixmybmc before an image that it is included on gets rolled out to the fleet.

note: this is not included in the image

Test Plan: run sync command on new host, run `fixmybmc` on that host

Reviewed By: alandau

Differential Revision: D63133410

fbshipit-source-id: 6cc2c17f423c9979001ffda04f84d9ccc13183e5