gtartemis: Add SEL when we do not discover expected number of modules
Summary:
As title, after probe check if we discovered 24 ASIC modules.
If we did not, raise a SEL.

Test Plan:
Tested shell snippet on a live artemis BMC and ensure we get the the
success info log. Change the threshold to 25 devices in the test
script and ensure that we wait 5s and print the error SEL.

Flash on GTArtemis and ensure we get a user.info log:
```
 2024 Oct 18 13:35:22 $BMC_HOST user.info gtartemis-54f9e98ca6: usbmon: FRU: 16 Detected 24 USB devices
```
Change 24 to 25 in code and flash on gtartemis and check log-util, we should see a SEL for CB.
```
root@rtptest165-oob:~# log-util cb --print
16   cb       2024-10-18 13:55:37    usbmon           FRU: 16 ASSERT: Detected 24 USB devices instead of expected 25
```

Reviewed By: williamspatrick

Differential Revision: D64619453

fbshipit-source-id: b3e62ce6c5d1d093763d46e4a03fceef6cd0f493