linux-6.6: apply mctp patch for hangs
Summary:
We are seeing with the upstream sync that includes
6.6.92 an issue on some platforms where mctpd will crash and
mctp-cli calls can hang.  This was identified as a bug in the
kernel and a fix was sent upstream.  Backport this fix while
it works through the kernel and stable-backporting process
upstream.

Test Plan:
Tested in QEMU against upstream lf-openbmc.  Harma
no longer hangs in boot-up.

Rollback Plan:

Reviewed By: cooperlees

Differential Revision: D76126484

fbshipit-source-id: 80f1ac62948087c7c3e6da2ff2bf1f360ad0a810