yosemite4: state: Block power cycle when BMC is not ready
Summary:
# Description:
Enable the only-allow-boot-when-bmc-ready option in phosphor-state-manager.

# Motivation:
For JIRA-1876, performing a blade chassis power cycle when the BMC is not ready can cause the blade host power to fail to turn on. Therefore, we need to prevent power cycle operations when the BMC is not ready.

X-link: https://github.com/facebookexternal/openbmc.wiwynn/pull/3329

Test Plan:
Build code – pass
Use mfg-tool to perform a blade chassis power cycle while BMC is not ready and confirm the operation is blocked – pass

Reviewed By: wangx6f

Differential Revision: D76257557

fbshipit-source-id: e188c5e4d75eb3aa68eb6c4dffc89efa998ef6f8