psu: support gpio bridge
Summary:
Support for GPIO bridge is required for certain components that need a GPIO to be set before reading, such as battery voltage. To obtain a valid reading, the BridgeGpio option can be configured in the Entity Manager JSON, similar to the implementation used for ADC sensors.

link:80418: psu: support gpio bridge
https://gerrit.openbmc.org/c/openbmc/dbus-sensors/+/80418

X-link: https://github.com/facebookexternal/openbmc.quanta/pull/4679

Reviewed By: ifel

Differential Revision: D76958276

fbshipit-source-id: 673566ab6fefb8e505611d0fc303edb5e96a910e