redfish-client: add systemd watchdog support
Summary:
We've seen at least one case where the redfish-client was
hung on a Catalina system.  I introduced a feature upstream in
sdbusplus[1] to enable automatic support for the systemd `WatchdogSec`
feature, which requires a daemon to pet a watchdog periodically for
systemd to know that it is still alive and functional.  Enable this
feature by setting the `WatchdogSec`.

[1]: https://gerrit.openbmc.org/c/openbmc/sdbusplus/+/81848

Test Plan:
Ran as-is and observed systemd killing the redfish-client.
Applied upstream code as a patch and observed normal behavior is
resumed (indicating that the watchdog is being petted).

Rollback Plan:

Reviewed By: ifel, jamesatha

Differential Revision: D78139441

fbshipit-source-id: 82e80573d7fa59b5b76be940dbb56b1719a2c4ad