MI300: Ignore GPU_RETIMER_*_1V2_VR_TEMP when considering sensor failure
Summary:
Sensor GPU_RETIMER_4_5_6_7_1V2_VR_TEMP and GPU_RETIMER_4_5_6_7_1V2_VR_TEMP
are applicable only for MI355. These sensors are NA on MI300.
Hence, on MI300, sensord was incorrectly stopping fscd and putting it in
boost-mode.

The correct fix should be to create a new sensor list for MI355 in
contrast with MI300. But, as a temporary workaround, just ignore
these two sensors in determining sensor-failure while we work on
the proper mechanism.

Test Plan:
Build libpal and load it onto the problematic BMC and restart sensord.
Start fscd manually and verify that sensord does not stop it.

Rollback Plan:

Reviewed By: jamesatha

Differential Revision:
D80215908

Privacy Context Container: L1188897

fbshipit-source-id: db544b65acf91741c017ee73e7693be593390a48