Ventura: Enable I2C slave timeout
Summary:
Enable the I2C slave timeout property in the Harma device tree to prevent the slave interface from hanging due to incomplete or abnormal I2C transactions, such as those occurring during MMC reboot.

This is a downstream patch that needs to wait for the driver patch merge on the upstream. The latest version of the driver patch lore link is as follows:
https://lore.kernel.org/all/20250820051832.3605405-1-ryan_chen@aspeedtech.com/

X-link: https://github.com/facebookexternal/openbmc.quanta/pull/4761

Reviewed By: amithash

Differential Revision: D80795531

fbshipit-source-id: 1c56acc9777f5ab24a7896a0ea23252eedc91e26