clemente: fix fan sensor name to adhere to RF naming conventions
Summary:
# Description

This change resolves the issue where sensor names don't match the RF
naming convention and modifies the sensor names in the Entity-Manager
configuration files for nvidia_gb300 to adhere to the naming rule.

This modification has already been committed to the upstream[1].
Since this modification concerns NVIDIA, it requires more time for
discussions with NVIDIA personnel and their review. Therefore,
this fix has been temporarily added to meet project requirements.

[1]:https://gerrit.openbmc.org/c/openbmc/entity-manager/+/83383

# Motivation

All sensor names need to adhere to the RF naming conventions.

X-link: https://github.com/facebookexternal/openbmc.ingrasys/pull/45

Test Plan:
Use the mfg-tool to check if the sensor names have been
successfully modified.

`mfg-tool sensor-display 2>/dev/null | table-sensor-display`

FAN0_PWM_PCT ok 80 Percent N/A N/A N/A N/A N/A N/A
FAN1_PWM_PCT ok 80 Percent N/A N/A N/A N/A N/A N/A
FAN2_PWM_PCT ok 80 Percent N/A N/A N/A N/A N/A N/A
FAN3_PWM_PCT ok 80 Percent N/A N/A N/A N/A N/A N/A
FAN4_PWM_PCT ok 80 Percent N/A N/A N/A N/A N/A N/A
FAN5_PWM_PCT ok 80 Percent N/A N/A N/A N/A N/A N/A
FAN6_PWM_PCT ok 80 Percent N/A N/A N/A N/A N/A N/A
FAN7_PWM_PCT ok 80 Percent N/A N/A N/A N/A N/A N/A

Reviewed By: williamspatrick

Differential Revision: D82088893

fbshipit-source-id: d21d2ed69cfc1458232bf06e916c349b5d6ec224