net: ncsi: check for netlink-driven responses before requiring a handler
Summary:
Move this patch to common.

Lore link:
https://lore.kernel.org/netdev/7346cdea74daedd0abecfcdcba2e7aa3be203a6e.camel@codeconstruct.com.au/T/#t

Currently, the NCSI response path will require an opcode-specific handler for all incoming messages. However, we may be receiving a response from a netlink-generated request, which may not have a corresponding in-kernel handler for that request opcode.

Check for the pending request (and hence for NETLINK_DRIVEN) before requiring a handler, deferring the check until we know it's a kernel-driven command.

X-link: https://github.com/facebookexternal/openbmc.quanta/pull/4809

Reviewed By: williamspatrick

Differential Revision: D85138712

fbshipit-source-id: f8f0a7019fa13bb2141596403f15f7d6e21a1edb