ptest-runner: remove tmpfs option
Summary:
The tmpfs option isn't used anywhere and likely is unworking
deadcode.  There is a lot of content we expect to be working in a
sandbox for running the ptests, which isn't started in the tmpfs path.

Delete this option to simplify the code.

Test Plan: Dead code; check CIT.

Reviewed By: amithash

Differential Revision:
D86135543

Privacy Context Container: L1188897

fbshipit-source-id: ed4e0dccd0ded20cadeadd446724c50299f04455