Add usb serial monitor to request for a managed reboot on error
Summary: This adds a service which runs every 3min, detects if the BMC is in the particular issue. If so, tries to remediate by reseting the USB host controller and restarting rackmon. If that fails to remediate, we will call managed-reboot which will create an event log requesting for a managed reboot.

Test Plan:
Tested out on a system with a FTDI chip, ensure the script is a no-op.
Tested it out on a GB200. Ensure its a no-op.
Tested it out on a system exhibiting the issue (By running the script directly)
```
root@rmc033-oob:/tmp# ./usb-serial-monitor.sh
Detected 268 checking it
terminate called after throwing an instance of 'std::system_error'
  what():  recv: Resource temporarily unavailable
./usb-serial-monitor.sh: line 3: 31650 Aborted                 (core dumped) rackmoncli read "$addr" 0x9000 > /dev/null
Device 268 failed
Detected 524 checking it
terminate called after throwing an instance of 'std::system_error'
  what():  recv: Resource temporarily unavailable
./usb-serial-monitor.sh: line 3:  3396 Aborted                 (core dumped) rackmoncli read "$addr" 0x9000 > /dev/null
Device 524 failed
Detected 780 checking it
terminate called after throwing an instance of 'std::system_error'
  what():  recv: Resource temporarily unavailable
./usb-serial-monitor.sh: line 3:  7321 Aborted                 (core dumped) rackmoncli read "$addr" 0x9000 > /dev/null
Device 780 failed
./usb-serial-monitor.sh: line 50: echo: write error: No such device
Detected 268 checking it
Detected 524 checking it
Detected 780 checking it
Remediated on 1 try
Auto-remediation successful!
```
Flash a BMC and ensure the service is running and checking the device.

Reviewed By: jamesatha

Differential Revision:
D85876275

Privacy Context Container: L1188897

fbshipit-source-id: 864a8386dec5e4864af3b4548b60f353b2a652ce