dbus-sensors: move PSUSensor 0204 patch to global SRC_URI
Summary:
Don't know why the patch 0204-PSUSensor-add-property-to-sensorTable-and-labelMatch.patch was previously applied only for fb-compute-multihost, as it is not specific to multihost platforms. Move it from the fb-compute-multihost override to the global SRC_URI so that it applies to all platforms.

Gerrit link of this patch: https://gerrit.openbmc.org/c/openbmc/dbus-sensors/+/83634

X-link: https://github.com/facebookexternal/openbmc.quanta/pull/4818

Test Plan: Verified successful reading of RTQ6056 sensors on Yv5.

Reviewed By: jagpalgill

Differential Revision: D86412034

fbshipit-source-id: 43193a03fcaad4bb1cd38ba4c8aabd370721cfe5