clemente: add compute-software-id for pldm fw update
Summary:
# Description

Add compute-software-id to the image. This is used in pldm-fw-update.sh from Yosemite4 to confirm the processing update package is the same as the one
we uploaded. Clemente will use similar approach to update firmware of the MCU CX8 IO Mezzanine via
PLDM.

# Motivation

We need to check if the processing package is the correct one during MCU PLDM firmware update.

X-link: https://github.com/facebookexternal/openbmc.ingrasys/pull/64

Test Plan: The command works in a downsized viersion of pldm-fw-update.sh that is used to update CX8 IO Mezzanine MCU firmware on Clemente.

Reviewed By: wangx6f

Differential Revision: D87034755

fbshipit-source-id: 86ee160f98d9b62225a7cef0f85ad4953ed6dae0