configurations: nvidia: Use external senosrs for gb300
Summary:
# Description

The only way to access all sensors on a CX8 is through PLDM. To do so we need to expose them as external sensors in order to update their values with those obtained by PLDM command.
This is associated with
85250: meta-facebook:clemente: enable mctp over usb for cx8 | https://gerrit.openbmc.org/c/openbmc/openbmc/+/85250

# Motivation

To access all sensors of CX8, we need to go with PLDM. The extrenal sensor allows us to set the value with the one retrieved through PLDM.

X-link: https://github.com/facebookexternal/openbmc.ingrasys/pull/63

Test Plan: Sensor values can be updated with the value obtained by pldmtool.

Reviewed By: wangx6f

Differential Revision: D87034607

fbshipit-source-id: 5640e05f5c9f6ce2dda3ed0a1fcd059c7a503e9f