clemente:net:mctp: add gateway routing support
Summary:
# Description

This change is from 6.12 (commit ad39c12) and is crucial for MCTP bridge to work. Given the kernal code will not be updated to where we need it to be soon. A patch to enable the feature is required to enable the MCU on NVIDIA CX8 IO Mezz to function as a MCTP bridge.
This is associated with
85250: meta-facebook:clemente: enable mctp over usb for cx8 | https://gerrit.openbmc.org/c/openbmc/openbmc/+/85250

# Motivation

We need to enable MCTP over USB for CX8 to avoid the complex MUX switching on the I2C side.

X-link: https://github.com/facebookexternal/openbmc.ingrasys/pull/62

Test Plan: Tested with the fix and MCU can finally bridging the message and no more "mctpd: Failed to add gateway route for EID #: Invalid argument" errors.

Reviewed By: wangx6f

Differential Revision: D87034384

fbshipit-source-id: 523dada48c0e7f03483808e57692c7ac6aa7ff96