santabarbara: adjust PLDM sensor polling and recovery
Summary:
- Change PLDM sensor polling interval to 2s
- Disable Dmctp recovery

The BMC is too busy with the current sensor polling rate, so we reduce the polling frequency by increasing the polling interval to 2 seconds.

When the MMC is busy, PLDM may remove the EID. Disabling Dmctp recovery helps keep the PLDM connection and allows it to recover when the MMC comes back.

X-link: https://github.com/facebookexternal/openbmc.quanta/pull/4859

Reviewed By: amithash

Differential Revision: D88263160

fbshipit-source-id: d299727340bf86d26a8659c158c67175ec20a6e0