rsyslog: Make remote logging more reliable
Summary:
In order of most impactful to least impactful:
(1) Change the queue type from linked list to fixed array. This
should make it perform better. Considering the change only allocates
the nodes (not the space for the messages) this is pretty harmless
especially considering the queue sizes we are working with.
(2) Queue size. Bumped it up from 100 to 10000. This should allow
the BMC to better handle the flood of console logs you tend to get
during power on. 100 Was abysmally small.
(3) discardseverity was bumped from (0) to ( 8 ). Before it would
discard all messages, now it would discard no messages. It should
prevent unintentional message loss
(4) timeoutenqueue was bumped from 0 (Do not wait) to 100ms. This
should add some push back.
(5) timeoutshutdown was bumped from 10ms to 100ms to match with
timeoutenqueue.
(6) reportSuspension was set to on. So we may debug issues with missing
targets (If it is really happening in the wild)
(7) workerthreadminimummessages removed config. If we are using only
one worker, this config seems pointless.
( 8 ) reorganized the options so all the queues and action configs
are together.

Test Plan:
Tested on Yosemite4 to ensure there are no obvious errors and the memory of rsyslog is managable.
```
6362     1 root     S    97992  11%   0   0% /usr/sbin/rsyslogd -n -i/var/run/rsyslogd.pid
```

Reviewed By: waffle2k

Differential Revision: D79205060

fbshipit-source-id: 5d318189d3dc595d5003599dea9f4fdfe44861fc